fiks: minor detail
