--- app/admin/artikel/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function page() {
-    return (
-        <div>
-            <h1>lorem ipsum dolor sit cihuy</h1>
-        </div>
-    )
-}
