+++ app/admin/dashboard/quiz/page.tsx
@@ -27,15 +27,23 @@ const QuizPage: FC = () => {
         router.push(`quiz/${quizId}`)
     }
 
+
     return (
         <div className="px-6 max-w-[80rem] mx-auto">
-            <header className="flex flex-col gap-4 py-4">
-                <h3 className="font-semibold text-3xl ">Quiz</h3>
-                <div className="flex justify-end">
-                    <Button onClick={() => router.push(`quiz/create`)} className="w-[5rem] sm:w-[10rem]">Add</Button>
+            <h3 className="font-semibold text-3xl ">Quiz</h3>
+            <p>
+                Total Quiz : {QUIZZES.length}
+            </p>
+                <div className="flex justify-between items-center mb-4 mt-4">
+                    <header className="flex flex-col gap-4 py-4">
+                        <div className="flex justify-end">
+                            <Button onClick={() => router.push(`quiz/create`)} className="w-[5rem] sm:w-[10rem]">Add</Button>
+                        </div>
+                    </header>
+                <div>
+                    tombol search
                 </div>
-            </header>
-
+            </div>
             <Table>
                 <TableHeader>
                     <TableRow>
@@ -63,12 +71,6 @@ const QuizPage: FC = () => {
                         </TableRow>
                     ))}
                 </TableBody>
-                <TableFooter>
-                    <TableRow>
-                        <TableCell colSpan={3}>Total:</TableCell>
-                        <TableCell className="text-center">{QUIZZES.length}</TableCell>
-                    </TableRow>
-                </TableFooter>
             </Table>
         </div>
     )
